feat: add reporting capability (#104)
This commit adds the "Reporter" trait. Users can pass implementations of
this trait to `create_graph` which will then call the `on_load` method
on the reporter for each module load. This can be used to display a
progress indicator of the graph building.

Reporting is not exposed in the JS API right now.