feat: add support for @ts-self-types pragma
This commit adds a `@ts-self-types="..."` pragma that
may be placed at the top of untyped code files,
such as JS and JSX files. The specifier in this
pragma will then be used as the types for this
file when it is imported into a different file.

This is effectively the same as the existing
`/// <reference types="..." />` support in untyped
files, but with the added benefit of not
overloading the TypeScript triple-slash directive
syntax, and working with block comments.

Unlike `@deno-types`, quotes are required around
the specifier.