fix: handle return statements in functions correctly
We now correctly make a distinction between expression like and declaration like functions when inferring return types based on return statements.

The biggest change here is that expression like functions that have no `return` statements in the function body will now not infer to `void` anymore. This is because in some cases TypeScript will instead infer the return type to `never`. This applies to function expressions, arrow functions, and object literal function expression shorthand. It does not apply to function declarations and class members.